refactor(tui): extract workspace tree owner and domain root (W1)
Create src/tui/workspace/ as the Workspace domain: paths + tree-data are
equivalent moves of workspace-paths/workspace-files, tree-owner.ts owns the
tree and quick-open state with narrow onOpenPath/onLoadError ports, and
controller.ts keeps page/focus composition plus the remaining domains until
their own waves. Root workspace-controller.ts is now a thin compat re-export
to be deleted in W5; no behavior change.
