+++ src/tui/workspace/paths.ts
